+++ include/bitcoin/bitcoin/message/announce_version.hpp
@@ -82,7 +82,7 @@ public:
 
     void reset();
 
-    size_t satoshi_size() const;
+    uint64_t satoshi_size() const;
 
 private:
